pip_audit, test: generalize caching middleware
